IShutdownDelegatorにProcess境界から実装を与える
- 今まで通りasync-exit-hookをラップしてGraceful Shutdown
- childprocess.send('shutdown') は今まで通り使える
